gfx/Projection: Move PerspectiveProjection here
